Graph types

All graphs which can be used with the sampling algorithms belong to a type hierarchy. At the top of the hierarchy, there is AbstractGraph:

RRRMC.Interface.AbstractGraph — Type.

AbstractGraph{ET<:Real}

An abstract type representing an Ising spin model. The ET parameter is the type returned by the energy and delta_energy functions.

See also SimpleGraph, DiscrGraph and DoubleGraph.

source

There are currently three abstract subclasses, which determine which sampling algorithms can be used:

RRRMC.Interface.SimpleGraph — Type.

SimpleGraph{ET} <: AbstractGraph{ET}

An abstract type representing a generic graph. This can only be used with standardMC, not with rrrMC or bklMC.

The ET parameter is the type returned by energy and delta_energy.

source

RRRMC.Interface.DiscrGraph — Type.

DiscrGraph{ET} <: AbstractGraph{ET}

An abstract type representing a graph in which the delta_energy values produced when flipping a spin belong to a finite discrete set, and thus can be sampled efficiently with rrrMC or bklMC.

The ET parameter is the type returned by energy and delta_energy.

It is also used internally in DoubleGraph.

See also neighbors and allΔE.

source

RRRMC.Interface.DoubleGraph — Type.

DoubleGraph{ET} <: AbstractGraph{ET}

An abstract type representing a graph in which the energy is the sum of two contributions, one of which can be encoded in a DiscrGraph type. This allows rrrMC to sample values more efficiently.

The ET parameter is the type returned by the energy and delta_energy functions. Note that it can be different from the type of the internal DiscrGraph object (e.g., one can have a DiscrGraph{Int} object inside a DoubleGraph{Float64} object).

See also discr_graph, delta_energy_residual and update_cache_residual!.

source