Built-in graphs

Following is the list of the graph models which are provided with the module. After loading the RRRMC module, they can be constructed like in this example:

julia> X = RRRMC.GraphRRG(10, 3)

Note that for models which involve randomness in the constructor you may want to set the random seed with srand before calling the constructor, for reproducibility purposes.

Spin glass models

Random regular graphs

RRRMC.RRG.GraphRRG — Type.

GraphRRG(N::Integer, K::Integer, LEV = (-1,1)) <: DiscrGraph

A DiscGraph implementing a random regular graph with N spins and connectivity K. Note: N*K must be even. Also, the graph generator uses the pairing model method by Bollobás, with a cutoff on the number of restarts, and thus it may occasionally fail if K is large. The interactions are extracted at random from LEV, which must be a Tuple of Reals. No external fields.

RRRMC.RRG.GraphRRGCont — Type.

GraphRRGCont(N::Integer, K::Integer, LEV) <: DoubleGraph{Float64}

A DoubleGraph implementing a random regular graph with N spins and connectivity K. Note: N*K must be even. Also, the graph generator uses the pairing model method by Bollobás, with a cutoff on the number of restarts, and thus it may occasionally fail if K is large. The interactions are extracted from a normal distribution with unit variance, and are then discretized using the values in LEV, which must be a Tuple of Reals. No external fields.

Same as GraphRRGContSimple, but it can be used with rrrMC.

RRRMC.RRG.GraphRRGContSimple — Type.

GraphRRGContSimple(N::Integer, K::Integer) <: SimpleGraph{Flaot64}

A SimpleGraph implementing a random regular graph with N spins and connectivity K. Note: N*K must be even. Also, the graph generator uses the pairing model method by Bollobás, with a cutoff on the number of restarts, and thus it may occasionally fail if K is large. The interactions are extracted from a normal distribution with unit variance.

Same as GraphRRGCont, but it's more efficient when used with standardMC.

Edwards-Anderson graphs

RRRMC.EA.GraphEA — Type.

GraphEA(L::Integer, D::Integer, LEV = (-1,1)) <: DiscrGraph

An Edwards-Anderson DiscrGraph: spins are arranged on a square lattice of size L in D dimensions (i.e. there are $L^D$ total spins), with periodic boundary conditions. The interactions are extracted at random from LEV, which must be a Tuple of Reals. No external fields.

RRRMC.EA.GraphEACont — Type.

GraphEACont(L::Integer, D::Integer, LEV) <: DoubleGraph{Float64}

An Edwards-Anderson DoubleGraph: spins are arranged on a square lattice of size L in D dimensions (i.e. there are $L^D$ total spins), with periodic boundary conditions. The interactions are extracted at random from a normal distribution with unit variance, and are then discretized using the values in LEV, which must be a Tuple of Reals. No external fields.

Same as GraphEAContSimple, but it can be used with rrrMC.

RRRMC.EA.GraphEAContSimple — Type.

GraphEACont(L::Integer, D::Integer) <: SimpleGraph{Float64}

An Edwards-Anderson SimpleGraph: spins are arranged on a square lattice of size L in D dimensions (i.e. there are $L^D$ total spins), with periodic boundary conditions. The interactions are extracted at random from a normal distribution with unit variance.

Same as GraphEACont, but it's more efficient when used with standardMC.

p-spin

RRRMC.PSpin3.GraphPSpin3 — Type.

GraphPSpin3(N::Integer, K::Integer) <: DiscrGraph

A DiscrGraph implementing a p-spin regular graph with p=3. N is the number of spins, and must be divisible by 3; K is the connectivity. All interactions are set to $J=1$.

Quantum models with transverse fields

RRRMC.QIsingT.GraphQIsingT — Type.

GraphQIsingT(N::Integer, M::Integer, Γ::Float64, β::Float64) <: DoubleGraph

A DoubleGraph which implements a quantum Ising spin model in a transverse magnetic field, using the Suzuki-Trotter transformation. N is the number of spins, M the number of Suzuki-Trotter replicas, Γ the transverse field, β the inverse temperature. The graph is fully-connected, the interactions are random ($J ∈ {-1,1}$), there are no external longitudinal fields.

See also Qenergy.

Trivial models used for testing and debugging

RRRMC.TwoSpin.GraphTwoSpin — Type.

GraphTwoSpin() <: DiscrGraph

A trivial DiscrGraph type with 2 spins inteacting ferromagnetically ($J=1$), without fields.

Only useful for testing/debugging purposes.

RRRMC.ThreeSpin.GraphThreeSpin — Type.

GraphThreeSpin() <: DiscrGraph

A trivial DiscrGraph type with 3 spins, ferromagnetic interactions ($J=1$), no fields, and periodic boundary conditions.

Only useful for testing/debugging purposes.

RRRMC.Fields.GraphFields — Type.

GraphFields(N::Integer, LEV::Tuple = (1,)) <: DiscrGraph

A simple DiscrGraph type with N non-interacting variables, each of which is subject to a local field. The fields are extracted at random from LEV, which must be a Tuple of Reals.

Mostly useful for testing/debugging purposes.

RRRMC.Fields.GraphFieldsCont — Type.

GraphFieldsCont(N::Integer, LEV::Tuple) <: DoubleGraph

A simple DoubleGraph type with N non-interacting variables, each of which is subject to a local field. The fields are extracted independently from a normal distribution with unit variance, and then are discretized using the values in LEV, which must be a Tuple of Reals.

Mostly useful for testing/debugging purposes.

RRRMC.Ising1D.GraphIsing1D — Type.

GraphIsing1D(N::Integer) <: DiscrGraph

A simple 1-dimensional DiscrGraph type with N spins, antiferromagnetic interactions ($J=-1$), no fields, and periodic boundary conditions.

Mostly useful for testing/debugging purposes.

RRRMC.Q0T.GraphQ0T — Type.

GraphQ0T(N::Integer, M::Integer, Γ::Float64, β::Float64) <: DoubleGraph

A simple DoubleGraph which implements independent spins in a transverse magnetic field, using the Suzuki-Trotter transformation. N is the number of spins, M the number of Suzuki-Trotter replicas, Γ the transverse field, β the inverse temperature.

Intended for testing/debugging purposes.

See also Qenergy.