Sampling algorithms

RRRMC.standardMC — Function.

standardMC(X::AbstractGraph, β::Real, iters::Integer; keywords...)

Runs iters iterations of a standard Metropolis Monte Carlo algorithm on the given Ising spin model X, at inverse temperature β. Each spin flip attempt counts as an iteration.

Returns two objects: a vector of energies and the last configuration (see Config).

Possible keyord arguments are:

step: the interval of iterations with which to collect energies (for the returned result) and to call the hook function (see below). The default is 1, which is good for debugging but otherwise generally a bad idea.

seed: the random seed. The default is some arbitrary number.

C0: the initial configuration. The default is nothing, in which case it is initialized at random. Otherwise it can be a Config object. Passing the result of a previous run can be useful e.g. when implementing a simulated annealing protocol, or if the system has not equilibrated yet.

hook: a function to be executed after every step number of iterations (see above). It must take five arguments: the current iteration, the graph X, the current configuration, the number of accepted moves so far, and the current energy. Useful to collect data other than the energy, write to files ecc; you'd probably want to use a closure, see the example below. The default is a no-op.

Basic example:

julia> srand(76543); X = RRRMC.GraphPSpin3(3999, 5); β = 1.0;
julia> Es, C = standardMC(X, β, 100_000, step = 1_000);

Example of using the hook for collecting samples as the columns of a BitMatrix:

julia> iters = 100_000; step = 1_000; l = iters ÷ step; N = RRRMC.getN(X);
julia> Cs = BitArray(N, l); hook = (it, X, C, acc, E) -> (Cs[:,it÷step]=C.s);
julia> Es, C = standardMC(X, β, iters, step = step, hook = hook);

source

RRRMC.rrrMC — Function.

rrrMC(X::AbstractGraph, β::Real, iters::Integer; keywords...)

Same as standardMC, but uses the reduced-rejection-rate method. Each iteration takes moretime, but has a higher chance of being accepted, so fewer iterations overall should be needed normally. Whether this trade-off is convenient depends on the parameters and the details of the model.

The return values and the keyword arguments are the same as standardMC, see the usage examples for that function. Note however that this function can only be used with DiscrGraph or DoubleGraph models.

source

RRRMC.bklMC — Function.

bklMC(X::DiscrGraph, β::Real, iters::Integer; keywords...)

Same as standardMC, but uses the rejection-free method by Bortz, Kalos and Lebowitz. Each step takes more, but rejected moves are essentially free, since they are skipped entirely.

The return values and the keyword arguments are the same as standardMC, see the usage examples for that function. Note however that this function can only be used with DiscrGraph models.

Note that the number of iterations includes the rejected moves. This makes the results directly comparable with those of standardMC. It also means that increasing β at fixed iters will result in fewer steps being actually computed.

source

RRRMC.Interface.Config — Type.

Config(N::Integer)

The object storing the configuration for an Ising model. Although the spin values are $σ_i ∈ {-1,1}$, internally they are stored in a BitArray, in the type field s, so that to obtain the real value one needs to perform the transformation $σ_i = 2s_i - 1$.

source